Hoist common OSThread fields and getters into AbstractOSThread
Second subsystem in the cross-OS consistency audit. Every OSThread
implementation declared the same thread-attribute fields (threadId, name,
state, kernel/user/start/up time, priority, faults, context switches, start
address) with identical return-the-field getters. Hoist them into
AbstractOSThread as protected fields + getters (the AbstractOSProcess
field-hoist model), so each platform keeps only its own parsing:

- AbstractOSThread: gains the 12 protected fields + field-backed getters.
- Linux/AIX/Solaris/BSD/Windows/Mac OSThread: drop the fields + getters, keep
  their parsing/constructors (Linux keeps its os field, Solaris its memoized
  psinfo suppliers, Windows its getProcName + perf-counter updateAttributes,
  BSD its ps-row parsing).
- BsdOSThread's protected fields (already the full superset) moved up; its
  checkstyle VisibilityModifier suppression migrated to AbstractOSThread.

Net -279 lines. getName() is now uniformly null-safe (returns empty per the
interface contract) instead of each platform differing; no observable change
for valid threads.
