+++ oshi-common/src/main/java/oshi/software/common/os/unix/aix/AixOSThread.java
@@ -17,65 +17,10 @@ import oshi.software.os.OSProcess;
 @ThreadSafe
 public class AixOSThread extends AbstractOSThread {
 
-    private int threadId;
-    private OSProcess.State state = OSProcess.State.INVALID;
-    private long startMemoryAddress;
-    private long contextSwitches;
-    private long kernelTime;
-    private long userTime;
-    private long startTime;
-    private long upTime;
-    private int priority;
-
     public AixOSThread(int pid, int tid) {
         super(pid);
         this.threadId = tid;
         updateAttributes();
-    }
-
-    @Override
-    public int getThreadId() {
-        return this.threadId;
-    }
-
-    @Override
-    public OSProcess.State getState() {
-        return this.state;
-    }
-
-    @Override
-    public long getStartMemoryAddress() {
-        return this.startMemoryAddress;
-    }
-
-    @Override
-    public long getContextSwitches() {
-        return this.contextSwitches;
-    }
-
-    @Override
-    public long getKernelTime() {
-        return this.kernelTime;
-    }
-
-    @Override
-    public long getUserTime() {
-        return this.userTime;
-    }
-
-    @Override
-    public long getUpTime() {
-        return this.upTime;
-    }
-
-    @Override
-    public long getStartTime() {
-        return this.startTime;
-    }
-
-    @Override
-    public int getPriority() {
-        return this.priority;
     }
 
     @Override
