+++ oshi-common/src/main/java/oshi/software/common/os/windows/WindowsOSThread.java
@@ -16,24 +16,12 @@ import static oshi.software.os.OSProcess.State.WAITING;
 import oshi.annotation.concurrent.ThreadSafe;
 import oshi.driver.common.windows.registry.ThreadPerfCounterBlock;
 import oshi.software.common.AbstractOSThread;
-import oshi.software.os.OSProcess.State;
 
 /**
  * Common base class for Windows OS thread implementations.
  */
 @ThreadSafe
 public abstract class WindowsOSThread extends AbstractOSThread {
-
-    private final int threadId;
-    private String name;
-    private State state;
-    private long startMemoryAddress;
-    private long contextSwitches;
-    private long kernelTime;
-    private long userTime;
-    private long startTime;
-    private long upTime;
-    private int priority;
 
     /**
      * Constructor.
@@ -47,56 +35,6 @@ public abstract class WindowsOSThread extends AbstractOSThread {
         super(pid);
         this.threadId = tid;
         updateAttributes(procName, pcb);
-    }
-
-    @Override
-    public int getThreadId() {
-        return threadId;
-    }
-
-    @Override
-    public String getName() {
-        return name;
-    }
-
-    @Override
-    public State getState() {
-        return state;
-    }
-
-    @Override
-    public long getStartMemoryAddress() {
-        return startMemoryAddress;
-    }
-
-    @Override
-    public long getContextSwitches() {
-        return contextSwitches;
-    }
-
-    @Override
-    public long getKernelTime() {
-        return kernelTime;
-    }
-
-    @Override
-    public long getUserTime() {
-        return userTime;
-    }
-
-    @Override
-    public long getStartTime() {
-        return startTime;
-    }
-
-    @Override
-    public long getUpTime() {
-        return upTime;
-    }
-
-    @Override
-    public int getPriority() {
-        return this.priority;
     }
 
     /**
